Some deletions done
