+++ RoomListDemo/SupportingFiles/Constants/Constants.swift
@@ -9,14 +9,9 @@
 import Foundation
 import UIKit
 
-enum MyError: Error {
-    case FoundNil(String)
-}
-
 typealias DataExists = (_ dataIsThere: Bool) -> Void
 
 //MARK: Others
-//let KEY_WINDOW = UIApplication.shared.keyWindow //Sometimes it will not allow u to add any view on KeyWindow, where as APP_KEY_WINDOW will always allow u to do the same .
 let APP_KEY_WINDOW = UIWindow.key
 
 let APP_DELEGATE = AppDelegate.shared
@@ -34,10 +29,3 @@ let LOADING_VIEW_TAG = 123456789
 let LOADING_IMAGE_TAG = 987654321
 
 let BACKEND_ERROR = 9000
-
-
-//struct AlertType {
-//    static let Default = 0
-//    static let Confirmation = 1
-//}
-
