+++ RoomListDemo/SupportingFiles/Helpers/AllExtensions.swift
@@ -155,3 +155,17 @@ extension UIViewController {
            SHOW_TOAST("Connection established")
        }
 }
+
+//MARK: Extention for Array
+
+extension Array {
+    func indexesOf<T : Equatable>(object:T) -> [Int] {
+        var result: [Int] = []
+        for (index,obj) in self.enumerated() {
+            if obj as! T == object {
+                result.append(index)
+            }
+        }
+        return result
+    }
+}
