+++ RoomListDemo/SupportingFiles/Helpers/Utils.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import ObjectiveC
 import UIKit
 
 private var utils: Utils? = nil
